+++ activestorage/app/controllers/active_storage/representations/proxy_controller.rb
@@ -11,7 +11,7 @@ class ActiveStorage::Representations::ProxyController < ActiveStorage::Represent
 
   def show
     http_cache_forever public: true do
-      send_blob_stream @representation.image, disposition: params[:disposition]
+      send_blob_stream @representation, disposition: params[:disposition]
     end
   end
 end